+++ lib/logic/tools/images_and_files/binary2image.dart
