+++ lib/widgets/tools/images_and_files/binary2image.dart
@@ -1,7 +1,7 @@
 import 'dart:typed_data';
 import 'package:flutter/material.dart';
 import 'package:gc_wizard/i18n/app_localizations.dart';
-import 'package:gc_wizard/logic/tools/science_and_technology/binary2image.dart';
+import 'package:gc_wizard/logic/tools/images_and_files/binary2image.dart';
 import 'package:gc_wizard/widgets/common/base/gcw_iconbutton.dart';
 import 'package:gc_wizard/widgets/common/base/gcw_textfield.dart';
 import 'package:gc_wizard/widgets/common/gcw_default_output.dart';
@@ -76,8 +76,8 @@ class Binary2ImageState extends State<Binary2Image> {
 
   _exportFile(BuildContext context, Uint8List data) async {
     var value = await saveByteDataToFile(
-        data.buffer.asByteData(), DateFormat('yyyyMMdd_HHmmss').format(DateTime.now()) + ".png");
+        data.buffer.asByteData(), 'image_export' + DateFormat('yyyyMMdd_HHmmss').format(DateTime.now()) + '.png');
 
-    if (value != null) showExportedFileDialog(context, value['path'], fileType: ".png", contentWidget: Image.memory(_outData));
+    if (value != null) showExportedFileDialog(context, value['path'], fileType: '.png', contentWidget: Image.memory(_outData));
   }
 }
